pipline/requirements.txt
rasterio>=1.3.0
Pillow>=9.0.0
numpy>=1.21.0
PyYAML>=6.0
requests>=2.28.0
tqdm>=4.64.0

